Redesign Openstack provider settings step to better support different credential types
