+++ src/app/node-data/basic/provider/openstack/template.html
@@ -82,7 +82,7 @@ limitations under the License.
                 fxFlex>
     Allocate Floating IP
     <i *ngIf="isFloatingIPEnforced()"
-       class="km-icon-info km-pointer"
+       class="km-icon-info icon-info km-pointer"
        matTooltip="Floating IP usage is enforced by selected datacenter"></i>
   </mat-checkbox>
