+++ src/app/shared/services/cluster.service.ts
@@ -10,13 +10,14 @@
 // limitations under the License.
 
 import {EventEmitter, Injectable} from '@angular/core';
+import {CloudSpec, Cluster, ClusterType} from '@shared/entity/cluster';
+import {SSHKey} from '@shared/entity/ssh-key';
+import {NodeProvider} from '@shared/model/NodeProviderConstants';
 import * as _ from 'lodash';
-import {CloudSpec, Cluster, ClusterType} from '../entity/cluster';
-import {SSHKey} from '../entity/ssh-key';
-import {NodeProvider} from '../model/NodeProviderConstants';
 
 @Injectable()
 export class ClusterService {
+  private _admissionPluginsEntity: string[] = [];
   readonly providerChanges = new EventEmitter<NodeProvider>();
   readonly datacenterChanges = new EventEmitter<string>();
   readonly sshKeyChanges = new EventEmitter<SSHKey[]>();
@@ -25,8 +26,10 @@ export class ClusterService {
   readonly clusterTypeChanges = new EventEmitter<ClusterType>();
 
   private _cluster: Cluster = Cluster.newEmptyClusterEntity();
-  private _sshKeys: SSHKey[] = [];
-  private _admissionPluginsEntity: string[] = [];
+
+  get cluster(): Cluster {
+    return this._cluster;
+  }
 
   set cluster(cluster: Cluster) {
     if (
@@ -43,8 +46,23 @@ export class ClusterService {
     this.clusterChanges.emit(this._cluster);
   }
 
-  get cluster(): Cluster {
-    return this._cluster;
+  private _sshKeys: SSHKey[] = [];
+
+  get sshKeys(): SSHKey[] {
+    return this._sshKeys;
+  }
+
+  set sshKeys(keys: SSHKey[]) {
+    this._sshKeys = keys;
+    this.sshKeyChanges.emit(this._sshKeys);
+  }
+
+  get provider(): NodeProvider {
+    const clusterProviders = Object.values(NodeProvider)
+      .map(provider => (this._cluster.spec.cloud[provider] ? provider : undefined))
+      .filter(p => p !== undefined);
+
+    return clusterProviders.length > 0 ? clusterProviders[0] : NodeProvider.NONE;
   }
 
   set provider(provider: NodeProvider) {
@@ -58,16 +76,12 @@ export class ClusterService {
     } as Cluster;
 
     if (provider) {
-      this.providerChanges.emit(provider);
+      this.providerChanges.next(provider);
     }
   }
 
-  get provider(): NodeProvider {
-    const clusterProviders = Object.values(NodeProvider)
-      .map(provider => (this._cluster.spec.cloud[provider] ? provider : undefined))
-      .filter(p => p !== undefined);
-
-    return clusterProviders.length > 0 ? clusterProviders[0] : NodeProvider.NONE;
+  get datacenter(): string {
+    return this._cluster.spec.cloud.dc;
   }
 
   set datacenter(datacenter: string) {
@@ -80,12 +94,8 @@ export class ClusterService {
     } as Cluster;
 
     if (datacenter) {
-      this.datacenterChanges.emit(datacenter);
+      this.datacenterChanges.next(datacenter);
     }
-  }
-
-  get datacenter(): string {
-    return this._cluster.spec.cloud.dc;
   }
 
   set labels(labels: object) {
@@ -97,23 +107,8 @@ export class ClusterService {
     this._cluster.spec.podNodeSelectorAdmissionPluginConfig = config;
   }
 
-  set sshKeys(keys: SSHKey[]) {
-    this._sshKeys = keys;
-    this.sshKeyChanges.emit(this._sshKeys);
-  }
-
-  get sshKeys(): SSHKey[] {
-    return this._sshKeys;
-  }
-
-  set admissionPlugins(plugins: string[]) {
-    this._admissionPluginsEntity = plugins;
-    this._cluster.spec.admissionPlugins = plugins;
-    this.admissionPluginsChanges.emit(this._admissionPluginsEntity);
-  }
-
-  get admissionPlugins(): string[] {
-    return this._admissionPluginsEntity;
+  get clusterType(): ClusterType {
+    return this._cluster.type;
   }
 
   set clusterType(type: ClusterType) {
@@ -121,8 +116,14 @@ export class ClusterService {
     this.clusterTypeChanges.emit(type);
   }
 
-  get clusterType(): ClusterType {
-    return this._cluster.type;
+  get admissionPlugins(): string[] {
+    return this._admissionPluginsEntity;
+  }
+
+  set admissionPlugins(plugins: string[]) {
+    this._admissionPluginsEntity = plugins;
+    this._cluster.spec.admissionPlugins = plugins;
+    this.admissionPluginsChanges.emit(this._admissionPluginsEntity);
   }
 
   reset(): void {
