+++ src/app/wizard/step/provider-settings/provider/basic/openstack/application/template.html
@@ -1,0 +1,55 @@
+<!--
+Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+<form [formGroup]="form"
+      fxLayout="column"
+      fxLayoutGap="16px">
+  <mat-form-field fxFlex>
+    <mat-label>Application Credential ID</mat-label>
+    <input [formControlName]="Controls.ApplicationCredentialID"
+           [name]="Controls.ApplicationCredentialID"
+           matInput
+           autocomplete="off"
+           required>
+    <mat-error *ngIf="form.get(Controls.ApplicationCredentialID).hasError('required')">
+      Application Credential ID is <strong>required</strong>.
+    </mat-error>
+  </mat-form-field>
+
+  <mat-form-field fxFlex>
+    <mat-label>Application Credential Secret</mat-label>
+    <input [formControlName]="Controls.ApplicationCredentialSecret"
+           [name]="Controls.ApplicationCredentialSecret"
+           matInput
+           type="password"
+           autocomplete="off"
+           required>
+    <mat-error *ngIf="form.get(Controls.ApplicationCredentialSecret).hasError('required')">
+      Application Credential Secret is <strong>required</strong>.
+    </mat-error>
+  </mat-form-field>
+
+  <km-combobox #floatingIPPoolCombobox
+               [required]="isRequired(Controls.FloatingIPPool)"
+               [grouped]="false"
+               [options]="floatingIPPools"
+               [hint]="getHint(Controls.FloatingIPPool)"
+               [formControlName]="Controls.FloatingIPPool"
+               (changed)="onFloatingIPPoolChange($event)"
+               [label]="floatingIPPoolsLabel"
+               inputLabel="Select Floating IP Pool..."
+               filterBy="name">
+    <div *option="let floatingIPPool">
+      {{floatingIPPool.name}}
+    </div>
+  </km-combobox>
+</form>
