+++ src/app/wizard/step/provider-settings/provider/basic/openstack/default/template.html
@@ -1,0 +1,86 @@
+<!--
+Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+<form [formGroup]="form"
+      fxLayout="column"
+      fxLayoutGap="16px">
+  <mat-form-field fxFlex>
+    <mat-label>Username</mat-label>
+    <input [formControlName]="Controls.Username"
+           [name]="Controls.Username"
+           matInput
+           autocomplete="off"
+           required>
+    <mat-error *ngIf="form.get(Controls.Username).hasError('required')">
+      Username is <strong>required</strong>.
+    </mat-error>
+  </mat-form-field>
+
+  <mat-form-field fxFlex>
+    <mat-label>Password</mat-label>
+    <input [formControlName]="Controls.Password"
+           [name]="Controls.Password"
+           matInput
+           type="password"
+           autocomplete="off"
+           required>
+    <mat-error *ngIf="form.get(Controls.Password).hasError('required')">
+      Password is <strong>required</strong>.
+    </mat-error>
+  </mat-form-field>
+
+  <mat-form-field fxFlex>
+    <mat-label>{{projectsLabel}}</mat-label>
+    <input matInput
+           [formControlName]="Controls.Project"
+           [name]="Controls.Project"
+           [matAutocomplete]="autoProject"
+           [required]="isRequired(Controls.Project)">
+    <mat-autocomplete #autoProject="matAutocomplete">
+      <mat-option *ngFor="let project of projects"
+                  [value]="project.name">
+        {{project.name}}
+      </mat-option>
+    </mat-autocomplete>
+    <mat-hint>{{getHint(Controls.Project)}}</mat-hint>
+    <mat-error *ngIf="form.get(Controls.Project).hasError('required')">
+      The Project is <strong>required</strong>.
+    </mat-error>
+  </mat-form-field>
+
+  <mat-form-field fxFlex>
+    <mat-label>Project ID</mat-label>
+    <input [formControlName]="Controls.ProjectID"
+           [name]="Controls.ProjectID"
+           [required]="isRequired(Controls.ProjectID)"
+           matInput
+           autocomplete="off">
+    <mat-error *ngIf="form.get(Controls.ProjectID).hasError('required')">
+      Project ID is <strong>required</strong>.
+    </mat-error>
+  </mat-form-field>
+
+  <km-combobox #floatingIPPoolCombobox
+               [required]="isRequired(Controls.FloatingIPPool)"
+               [grouped]="false"
+               [options]="floatingIPPools"
+               [hint]="getHint(Controls.FloatingIPPool)"
+               [formControlName]="Controls.FloatingIPPool"
+               (changed)="onFloatingIPPoolChange($event)"
+               [label]="floatingIPPoolsLabel"
+               inputLabel="Select Floating IP Pool..."
+               filterBy="name">
+    <div *option="let floatingIPPool">
+      {{floatingIPPool.name}}
+    </div>
+  </km-combobox>
+</form>
