+++ src/app/wizard/step/provider-settings/provider/basic/openstack/style.scss
@@ -1,0 +1,44 @@
+// Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+@import 'variables';
+
+.km-wizard-openstack-button-toggle-group {
+  &.mat-button-toggle-group {
+    border-radius: $border-radius;
+    margin-bottom: 0;
+
+    &.km-toggle-subject-type {
+      margin-left: 30px;
+    }
+
+    .mat-button-toggle {
+      border-radius: 0;
+      margin-right: 0;
+      max-height: 40px;
+      min-height: 40px;
+
+      &:first-child {
+        border-bottom-left-radius: $border-radius;
+        border-top-left-radius: $border-radius;
+        max-width: initial;
+        min-width: 82px;
+      }
+
+      &:last-child {
+        border-bottom-right-radius: $border-radius;
+        border-top-right-radius: $border-radius;
+        max-width: initial;
+        min-width: 115px;
+      }
+    }
+  }
+}
