+++ src/app/wizard/step/provider-settings/provider/basic/openstack/template.html
@@ -32,99 +32,24 @@ limitations under the License.
     </mat-error>
   </mat-form-field>
 
-  <mat-form-field fxFlex>
-    <mat-label>Username</mat-label>
-    <input [formControlName]="Controls.Username"
-           [name]="Controls.Username"
-           matInput
-           autocomplete="off"
-           required>
-    <mat-error *ngIf="form.get(Controls.Username).hasError('required')">
-      Username is <strong>required</strong>.
-    </mat-error>
-  </mat-form-field>
+  <mat-button-toggle-group class="km-wizard-openstack-button-toggle-group"
+                           group="bindingType"
+                           (change)="changeView($event)">
+    <mat-button-toggle [value]="CredentialsType.Default"
+                       [checked]="credentialsType === CredentialsType.Default">
+      User Credentials
+    </mat-button-toggle>
+    <mat-button-toggle [value]="CredentialsType.Application"
+                       [checked]="credentialsType === CredentialsType.Application">
+      Application Credentials
+    </mat-button-toggle>
+  </mat-button-toggle-group>
 
-  <mat-form-field fxFlex>
-    <mat-label>Password</mat-label>
-    <input [formControlName]="Controls.Password"
-           [name]="Controls.Password"
-           matInput
-           type="password"
-           autocomplete="off"
-           required>
-    <mat-error *ngIf="form.get(Controls.Password).hasError('required')">
-      Password is <strong>required</strong>.
-    </mat-error>
-  </mat-form-field>
+  <ng-container *ngIf="credentialsType === CredentialsType.Default">
+    <km-wizard-openstack-provider-basic-default-credentials [formControl]="form.get(Controls.Credentials)"></km-wizard-openstack-provider-basic-default-credentials>
+  </ng-container>
 
-  <mat-form-field fxFlex>
-    <mat-label>Application Credential ID</mat-label>
-    <input [formControlName]="Controls.ApplicationCredentialID"
-           [name]="Controls.ApplicationCredentialID"
-           matInput
-           autocomplete="off"
-           required>
-    <mat-error *ngIf="form.get(Controls.ApplicationCredentialID).hasError('required')">
-      Application Credential ID is <strong>required</strong>.
-    </mat-error>
-  </mat-form-field>
-
-  <mat-form-field fxFlex>
-    <mat-label>Application Credential Secret</mat-label>
-    <input [formControlName]="Controls.ApplicationCredentialSecret"
-           [name]="Controls.ApplicationCredentialSecret"
-           matInput
-           type="password"
-           autocomplete="off"
-           required>
-    <mat-error *ngIf="form.get(Controls.ApplicationCredentialSecret).hasError('required')">
-      Application Credential Secret is <strong>required</strong>.
-    </mat-error>
-  </mat-form-field>
-
-  <mat-form-field fxFlex>
-    <mat-label>{{projectsLabel}}</mat-label>
-    <input matInput
-           [formControlName]="Controls.Project"
-           [name]="Controls.Project"
-           [matAutocomplete]="autoProject"
-           [required]="isRequired(Controls.Project)">
-    <mat-autocomplete #autoProject="matAutocomplete">
-      <mat-option *ngFor="let project of projects"
-                  [value]="project.name">
-        {{project.name}}
-      </mat-option>
-    </mat-autocomplete>
-    <mat-hint>{{getHint(Controls.Project)}}</mat-hint>
-    <mat-error *ngIf="form.get(Controls.Project).hasError('required')">
-      The Project is <strong>required</strong>.
-    </mat-error>
-  </mat-form-field>
-
-  <mat-form-field fxFlex>
-    <mat-label>Project ID</mat-label>
-    <input [formControlName]="Controls.ProjectID"
-           [name]="Controls.ProjectID"
-           [required]="isRequired(Controls.ProjectID)"
-           matInput
-           autocomplete="off">
-    <mat-error *ngIf="form.get(Controls.ProjectID).hasError('required')">
-      Project ID is <strong>required</strong>.
-    </mat-error>
-  </mat-form-field>
-
-  <km-combobox #floatingIPPoolCombobox
-               [required]="isRequired(Controls.FloatingIPPool)"
-               [grouped]="false"
-               [options]="floatingIPPools"
-               [hint]="getHint(Controls.FloatingIPPool)"
-               [formControlName]="Controls.FloatingIPPool"
-               (changed)="onFloatingIPPoolChange($event)"
-               [label]="floatingIPPoolsLabel"
-               inputLabel="Select Floating IP Pool..."
-               filterBy="name">
-    <div *option="let floatingIPPool">
-      {{floatingIPPool.name}}
-    </div>
-  </km-combobox>
+  <ng-container *ngIf="credentialsType === CredentialsType.Application">
+    <km-wizard-openstack-provider-basic-app-credentials [formControl]="form.get(Controls.Credentials)"></km-wizard-openstack-provider-basic-app-credentials>
+  </ng-container>
 </form>
