+++ src/app/wizard/step/provider-settings/provider/basic/openstack/theme.scss
@@ -9,13 +9,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-.km-custom-disk-size {
-  align-items: center;
-  display: flex;
-  height: 65px;
-}
-
-.icon-info {
-  align-self: center;
-  margin-left: 5px;
+@mixin theme-wizard-openstack-provider-basic($colors) {
+  .km-wizard-openstack-button-toggle-group {
+    &.mat-button-toggle-group {
+      .mat-button-toggle {
+        &:not(.mat-button-toggle-checked) {
+          background-color: map-get($colors, divider);
+        }
+      }
+    }
+  }
 }
