+++ src/app/wizard/step/provider-settings/provider/extended/openstack/default/template.html
@@ -1,0 +1,62 @@
+<!--
+Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+<form [formGroup]="form"
+      fxLayout="column"
+      fxLayoutGap="16px">
+  <km-combobox #securityGroupCombobox
+               [required]="false"
+               [grouped]="false"
+               [options]="securityGroups"
+               [hint]="getHint(Controls.SecurityGroup)"
+               [formControlName]="Controls.SecurityGroup"
+               (changed)="onSecurityGroupChange($event)"
+               [label]="securityGroupsLabel"
+               inputLabel="Select security group..."
+               filterBy="name">
+    <div *option="let securityGroup">
+      {{securityGroup.name}}
+    </div>
+  </km-combobox>
+
+  <km-combobox #networkCombobox
+               [required]="false"
+               [grouped]="false"
+               [options]="networks"
+               [hint]="getHint(Controls.Network)"
+               [formControlName]="Controls.Network"
+               (changed)="onNetworkChange($event)"
+               [label]="networksLabel"
+               inputLabel="Select network..."
+               filterBy="name">
+    <div *option="let network">
+      {{network.name}}
+    </div>
+  </km-combobox>
+
+  <km-combobox #subnetIDCombobox
+               [required]="false"
+               [grouped]="false"
+               [valueFormatter]="subnetIDDisplayName.bind(this)"
+               [options]="subnetIDs"
+               [hint]="getHint(Controls.SubnetID)"
+               [formControlName]="Controls.SubnetID"
+               (changed)="onSubnetIDChange($event)"
+               [label]="subnetIDsLabel"
+               inputLabel="Select subnet id..."
+               filterBy="name"
+               selectBy="id">
+    <div *option="let subnet">
+      {{subnet.name}} ({{subnet.id}})
+    </div>
+  </km-combobox>
+</form>
