+++ src/app/wizard/step/provider-settings/provider/extended/openstack/template.html
@@ -10,53 +10,15 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 -->
+
 <form [formGroup]="form"
       fxLayout="column"
       fxLayoutGap="16px">
-  <km-combobox #securityGroupCombobox
-               [required]="false"
-               [grouped]="false"
-               [options]="securityGroups"
-               [hint]="getHint(Controls.SecurityGroup)"
-               [formControlName]="Controls.SecurityGroup"
-               (changed)="onSecurityGroupChange($event)"
-               [label]="securityGroupsLabel"
-               inputLabel="Select security group..."
-               filterBy="name">
-    <div *option="let securityGroup">
-      {{securityGroup.name}}
-    </div>
-  </km-combobox>
+  <ng-container *ngIf="credentialsType === CredentialsType.Default">
+    <km-wizard-openstack-provider-extended-default-credentials [formControl]="form.get(Controls.Credentials)"></km-wizard-openstack-provider-extended-default-credentials>
+  </ng-container>
 
-  <km-combobox #networkCombobox
-               [required]="false"
-               [grouped]="false"
-               [options]="networks"
-               [hint]="getHint(Controls.Network)"
-               [formControlName]="Controls.Network"
-               (changed)="onNetworkChange($event)"
-               [label]="networksLabel"
-               inputLabel="Select network..."
-               filterBy="name">
-    <div *option="let network">
-      {{network.name}}
-    </div>
-  </km-combobox>
-
-  <km-combobox #subnetIDCombobox
-               [required]="false"
-               [grouped]="false"
-               [valueFormatter]="subnetIDDisplayName.bind(this)"
-               [options]="subnetIDs"
-               [hint]="getHint(Controls.SubnetID)"
-               [formControlName]="Controls.SubnetID"
-               (changed)="onSubnetIDChange($event)"
-               [label]="subnetIDsLabel"
-               inputLabel="Select subnet id..."
-               filterBy="name"
-               selectBy="id">
-    <div *option="let subnet">
-      {{subnet.name}} ({{subnet.id}})
-    </div>
-  </km-combobox>
+  <ng-container *ngIf="credentialsType === CredentialsType.Application">
+    <km-wizard-openstack-provider-extended-app-credentials [formControl]="form.get(Controls.Credentials)"></km-wizard-openstack-provider-extended-app-credentials>
+  </ng-container>
 </form>
