+++ src/assets/css/theme/_component.scss
@@ -31,6 +31,7 @@
 @import 'src/app/cluster/cluster-details/rbac/add-binding/theme';
 @import 'src/app/settings/admin/presets/theme';
 @import 'src/app/settings/admin/presets/edit-dialog/theme';
+@import 'src/app/wizard/step/provider-settings/provider/basic/openstack/theme';
 
 @mixin theme-components($colors) {
   @include theme-addon-list-component($colors);
@@ -55,4 +56,5 @@
   @include theme-preset-list-component($colors);
   @include theme-preset-edit-dialog-component($colors);
   @include theme-changelog-component($colors);
+  @include theme-wizard-openstack-provider-basic($colors);
 }
